+++ frontend/src/components/Courses/CourseUtilComponents.tsx
@@ -12,9 +12,7 @@ import {
 } from "@mui/material";
 import {
   Course,
-  Project,
   ProjectDetail,
-  apiHost,
   getIdFromLink,
   getNearestFutureDate,
 } from "./CourseUtils";
@@ -22,7 +20,6 @@ import { Link, useNavigate, useLocation } from "react-router-dom";
 import { useState, useEffect, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import debounce from "debounce";
-import { authenticatedFetch } from "../../utils/authenticated-fetch";
 
 /**
  * @param text - The text to be displayed
@@ -79,9 +76,10 @@ export function SearchBox({
  * @returns A component to display courses in horizontal scroller where each course is a card containing its name.
  */
 export function SideScrollableCourses({
-  courses,
+  courses,projects
 }: {
   courses: Course[];
+  projects: {[courseId: string]: ProjectDetail[];};
 }): JSX.Element {
   //const navigate = useNavigate();
   const location = useLocation();
@@ -101,9 +99,6 @@ export function SideScrollableCourses({
   const [teacherNameFilter, setTeacherNameFilter] = useState(
     initialTeacherNameFilter
   );
-  const [projects, setProjects] = useState<{
-    [courseId: string]: ProjectDetail[];
-  }>({});
 
   const debouncedHandleSearchChange = useMemo(
     () =>
@@ -149,55 +144,6 @@ export function SideScrollableCourses({
     const newTeacherNameFilter = event.target.value;
     setTeacherNameFilter(newTeacherNameFilter);
   };
-
-  useEffect(() => {
-    // Fetch projects for each course
-    const fetchProjects = async () => {
-      const projectPromises = courses.map((course) =>
-        authenticatedFetch(
-          `${apiHost}/projects?course_id=${getIdFromLink(course.course_id)}`
-        ).then((response) => response.json())
-      );
-
-      const projectResults = await Promise.all(projectPromises);
-      const projectsMap: { [courseId: string]: ProjectDetail[] } = {};
-
-      projectResults.forEach((result, index) => {
-        const detailProjectPromises = result.data.map(async (item: Project) => {
-          const projectRes = await authenticatedFetch(item.project_id);
-          if (projectRes.status !== 200) {
-            throw new Response("Failed to fetch project data", {
-              status: projectRes.status,
-            });
-          }
-          const projectJson = await projectRes.json();
-          const projectData = projectJson.data;
-          let projectDeadlines = [];
-          if (projectData.deadlines) {
-            projectDeadlines = projectData.deadlines.map(
-              ([description, dateString]: [string, string]) => ({
-                description,
-                date: new Date(dateString),
-              })
-            );
-          }
-          const project: ProjectDetail = {
-            ...item,
-            deadlines: projectDeadlines,
-          };
-          return project;
-        });
-        Promise.all(detailProjectPromises).then((projects) => {
-          projectsMap[getIdFromLink(courses[index].course_id)] = projects;
-          setProjects({ ...projectsMap });
-        });
-      });
-
-      setProjects(projectsMap);
-    };
-
-    fetchProjects();
-  }, [courses]);
 
   const filteredCourses = courses.filter(
     (course) =>
